Move exchange metadata related funcs to separate file
